rename go app
