+++ internal/app/grafana-multi-tenant-proxy/auth/oauth_auth.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 
 	"github.com/coreos/go-oidc"
-	"github.com/giantswarm/loki-multi-tenant-proxy/internal/pkg"
+	"github.com/giantswarm/grafana-multi-tenant-proxy/internal/pkg"
 	"go.uber.org/zap"
 )
